+++ test/fixture/nuxt.config.ts
@@ -11,10 +11,12 @@ const config: NuxtConfig = {
   components: {
     dirs: [
       '~/components',
+      { path: '~/components/global', global: true },
       { path: '~/components/multifile', extensions: ['vue'] },
       '~/non-existent',
       { path: '@/components/base', prefix: 'Base' },
-      { path: '@/components/icons', prefix: 'Icon', transpile: true /* Only for coverage purpose */ }
+      { path: '@/components/icons', prefix: 'Icon', transpile: true /* Only for coverage purpose */ },
+      { path: '@/theme/components', level: 1 }
     ]
   },