+++ test/fixture/pages/index.vue
@@ -1,10 +1,12 @@
 <template>
   <div>
+    <Header />
     <Foo />
     <LazyBar />
     <BaseButton />
     <IconHome />
     <MAwesome />
     <Functional />
+    <PAScal />
   </div>
 </template>