+++ test/fixture/theme/components/Header.vue
@@ -1,0 +1,3 @@
+<template>
+  <header>theme header</header>
+</template>