+++ test/unit/tagExtractor.test.ts
@@ -1,10 +1,10 @@
-import path from 'path'
+import path from 'upath'
 import { extractTags } from '../../src/tagExtractor'
 
 test('with template', async () => {
   const tags = await extractTags(path.resolve('test/fixture/pages/index.vue'))
 
-  expect(tags).toEqual(['Foo', 'LazyBar', 'BaseButton', 'IconHome', 'MAwesome', 'Functional', 'div'])
+  expect(tags).toEqual(['Header', 'Foo', 'LazyBar', 'BaseButton', 'IconHome', 'MAwesome', 'Functional', 'PAScal', 'div'])
 })
 
 test('without template', async () => {